F5/requirements.txt
setuptools
f5-sdk
python-dateutil
